+++ app/src/main/java/com/tuneitall/tuner/Support.kt
@@ -1,0 +1,3 @@
+package com.tuneitall.tuner
+
+internal const val SUPPORT_URL = "https://www.buymeacoffee.com/majkey"