+++ app/src/test/java/com/tuneitall/tuner/SupportLinkTest.kt
@@ -1,0 +1,11 @@
+package com.tuneitall.tuner
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class SupportLinkTest {
+    @Test
+    fun supportUsesTheSamePublishedBuyMeACoffeePageAsScanIt() {
+        assertEquals("https://www.buymeacoffee.com/majkey", SUPPORT_URL)
+    }
+}